GeoBloc Server: Nearly complete internal overhaul; the server now works with packages, meaning you can send any type of file through the multipart post method. The files are stored as Blob data types inside a BasicPackageForm. Two new servlets were created for displaying the contents of the BasicPackageForms and for downloading them. Also, main.jsp was changed to work with BasicPackageForms instead of BasicForms, and of course, the servlet calls have changed. Also, a new servlet for clearing the database of BasicPackageForms was added.
Old code (old main.jsp, old BasicForms servlets) are still part of the source code, but not used. Will clean them later.

NOTE: Android client will not work with the new system, it will access the old servlets. It only needs to change a line of code; will work on it now.