GeoBloc (Android Client): So many changes I think I'll forget them... I changed IUploadPackageTaskListener to IStandardTaskListener, and changed the method called when the task is complete in order to receive a Java Object; this way the task caller is responsible of knowing what kind of Java Object it's going to receive. Added a new Activity, FormsDownloader, and a new task, GetListOfAvailableFormsTask, to fetch a list of form files (and their respective keys) from the server. Also built a new class, SimpleHttpPost, for the necessary doPost() method. Utilities know has a second method, for displaying a simple Dialog with custom text & title.
Outline: FormsDownloader fetches a list of forms stored in the server. No form version detection has been implemented, but it has been thought about in the UI (easy to see). Still missing the second half, which processes the available forms list to disable the already downloaded ones before they're loaded in the ListView, and allows the user to select and download forms (new task required, servlet is built)
Also do not understand why I put an erase button in the Activity...

To-Do:
- Define form new version detection
- Hashtable processing (already downloaded forms, new versions) before loading as Listview
- Select & Download forms (maybe even a way of seeing available ones?)