Some hard work going on here. r163 appeared to work fine on 1.6 devices, because both of them had the instances/packages folder built, and this was causing one crash in my 2.2 device. Next, there were SQLite crashes all over the place, with FormsManager, InstanceManager and FormActivity all causing the databases to rebuild themselves. A combination of separating the tables into two distinct databases plus some re-wiring appears to have fixed this. A major fault was the fact that separating the tables into two databases caused Exceptions because we were trying to read both forms and instances from the same database. This has been fixed, and no SQL Exceptions appear in LogCat from my testing.
Another bug caused the UpdateFormsService to not update the form's version when the list was updated, plus it wouldn't detect properly when a more recent version of a form was available. I believe this has been fixed, but I need tests.

Lastly, the detection at download time of multiple forms with the same filename had a bug causing the forms with the same id to store their file under the same path, meaning only one form was actually stored, causing problems when trying to open one of the duplicated forms after one was deleted. I know, in the first place, this case shouldn't be contemplated because it's the server's responsibility to assure us clients that there's only one form with a given id, but that's not how it works right now. So much for protocols.

By the way, removing the application and re-installing it is recommended due to database schema changes. I'll try to unite them again, which will again require a complete wipe of the app in another revision.